cast to unsigned to remove warning (gcc)
svn path=/trunk/KDE/kdepim/; revision=809757

+++ folderdialogacltab.cpp
@@ -202,7 +202,7 @@ unsigned int KMail::ACLEntryDialog::permissions() const
 {
   QAbstractButton* button = mButtonGroup->checkedButton();
   if( !button )
-    return -1; // hm ?
+    return static_cast<unsigned int>(-1); // hm ?
   return mButtonGroup->id( button );
 }
